+++ simulator/src/com/md/mechevo/game/weapon/MineLauncher.java
@@ -14,7 +14,7 @@ public class MineLauncher extends Weapon {
 
 	@Override
 	public void fire(State state, Solid target) {
-		float angle = super.getAngleToTarget(target);
+		double angle = super.getAngleToTarget(target);
 		Projectile proj = new Mine(state.getNextId(), this.getPlayer().getPosition(), angle, this);
 		state.addProjectile(proj);
