+++ simulator/src/com/md/mechevo/game/weapon/Minigun.java
@@ -13,7 +13,7 @@ public class Minigun extends Weapon {
 	}
 
 	public void fire(State state, Solid target) {
-		float angle = super.getAngleToTarget(target);
+		double angle = super.getAngleToTarget(target);
 		Projectile proj =
 				new Bullet(state.getNextId(), this.getPlayer().getPosition(), angle, this);
 		state.addProjectile(proj);
